pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "glm-hybrid-ocr"
version = "0.1.0"
description = "Document processing system using GLM-OCR for OCR and Qwen for image captioning"
requires-python = ">=3.10"
dependencies = [
    "glmocr[selfhosted]",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "Pillow>=10.0.0",
    "pypdfium2>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/glm_hybrid_ocr"]
